ch10-cyber-ui/pyproject.toml
[project]
name = "chainlit-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chainlit==2.9.6",
    "pandas>=3.0.1",
    "plotly>=6.5.2",
    "python-dotenv>=1.2.1",
]
